Refactor StatsStorage. Create new files for tuple samples storage.
