+++ src/include/optimizer/stats/stats_storage.h
@@ -25,16 +25,9 @@ namespace peloton {
 
 namespace storage {
 class DataTable;
-class Tuple;
-}
-
-namespace catalog {
-class Schema;
 }
 
 namespace optimizer {
-
-#define SAMPLES_DB_NAME "samples_db"
 
 using ValueFrequencyPair = std::pair<type::Value, double>;
 
@@ -45,18 +38,16 @@ class StatsStorage {
 
   StatsStorage();
 
-  ~StatsStorage();
-
   /* Functions for managing stats table and schema */
 
   void CreateStatsCatalog();
 
   /* Functions for adding, updating and quering stats */
 
-  void AddOrUpdateTableStats(storage::DataTable *table, TableStats *table_stats,
+  void InsertOrUpdateTableStats(storage::DataTable *table, TableStats *table_stats,
                              concurrency::Transaction *txn = nullptr);
 
-  void AddOrUpdateColumnStats(oid_t database_id, oid_t table_id,
+  void InsertOrUpdateColumnStats(oid_t database_id, oid_t table_id,
                               oid_t column_id, int num_row, double cardinality,
                               double frac_null, std::string most_common_vals,
                               double most_common_freqs,
@@ -65,28 +56,6 @@ class StatsStorage {
 
   std::unique_ptr<std::vector<type::Value>> GetColumnStatsByID(
       oid_t database_id, oid_t table_id, oid_t column_id);
-
-  /* Functions for managing tuple samples */
-
-  void CreateSamplesDatabase();
-
-  void AddSamplesTable(
-      storage::DataTable *data_table,
-      std::vector<std::unique_ptr<storage::Tuple>> &sampled_tuples);
-
-  bool InsertSampleTuple(storage::DataTable *samples_table,
-                         std::unique_ptr<storage::Tuple> tuple,
-                         concurrency::Transaction *txn);
-
-  void GetTupleSamples(oid_t database_id, oid_t table_id,
-                       std::vector<storage::Tuple> &tuple_samples);
-
-  void GetColumnSamples(oid_t database_id, oid_t table_id, oid_t column_id,
-                        std::vector<type::Value> &column_samples);
-
-  std::string GenerateSamplesTableName(oid_t db_id, oid_t table_id) {
-    return std::to_string(db_id) + "_" + std::to_string(table_id);
-  }
 
   /* Functions for triggerring stats collection */
