+++ src/optimizer/stats/stats_storage.cpp
@@ -15,7 +15,6 @@
 #include "catalog/column_stats_catalog.h"
 #include "type/value.h"
 #include "storage/data_table.h"
-#include "storage/table_factory.h"
 
 namespace peloton {
 namespace optimizer {
@@ -33,21 +32,6 @@ StatsStorage *StatsStorage::GetInstance(void) {
 StatsStorage::StatsStorage() {
   pool_.reset(new type::EphemeralPool());
   CreateStatsCatalog();
-  CreateSamplesDatabase();
-}
-
-/**
- * ~StatsStorage - Deconstructor of StatsStorage.
- * It deletes(drops) the 'samples_db' from catalog in case of memory leak.
- * TODO: Remove this when catalog frees the databases memory in its
- * deconstructor
- * in the future.
- */
-StatsStorage::~StatsStorage() {
-  auto &txn_manager = concurrency::TransactionManagerFactory::GetInstance();
-  auto txn = txn_manager.BeginTransaction();
-  catalog::Catalog::GetInstance()->DropDatabaseWithName(SAMPLES_DB_NAME, txn);
-  txn_manager.CommitTransaction(txn);
 }
 
 /**
@@ -61,13 +45,11 @@ void StatsStorage::CreateStatsCatalog() {
 }
 
 /**
- * AddOrUpdateTableStats - Add or update all column stats of a table.
+ * InsertOrUpdateTableStats - Add or update all column stats of a table.
  * This function iterates all column stats in the table stats and insert column
  * stats tuples into the 'stats' table in the catalog database.
- * This function only add table stats to the catalog for now.
- * TODO: Implement stats UPDATE if the column stats already exists.
  */
-void StatsStorage::AddOrUpdateTableStats(storage::DataTable *table,
+void StatsStorage::InsertOrUpdateTableStats(storage::DataTable *table,
                                          TableStats *table_stats,
                                          concurrency::Transaction *txn) {
   // Add or update column stats sequentially.
@@ -97,13 +79,17 @@ void StatsStorage::AddOrUpdateTableStats(storage::DataTable *table,
       histogram_bounds_str = ConvertDoubleArrayToString(histogram_bounds);
     }
 
-    AddOrUpdateColumnStats(database_id, table_id, column_id, num_row,
+    LOG_DEBUG(
+        "InsertOrUpdateColumnStats: num_row: %lu, cardinality: %lf, frac_null: "
+        "%lf",
+        num_row, cardinality, frac_null);
+    InsertOrUpdateColumnStats(database_id, table_id, column_id, num_row,
                            cardinality, frac_null, most_common_val_str,
                            most_common_freq, histogram_bounds_str, txn);
   }
 }
 
-void StatsStorage::AddOrUpdateColumnStats(oid_t database_id, oid_t table_id,
+void StatsStorage::InsertOrUpdateColumnStats(oid_t database_id, oid_t table_id,
                                           oid_t column_id, int num_row,
                                           double cardinality, double frac_null,
                                           std::string most_common_vals,
@@ -131,7 +117,6 @@ void StatsStorage::AddOrUpdateColumnStats(oid_t database_id, oid_t table_id,
 
 /**
  * GetColumnStatsByID - Query the 'stats' table to get the column stats by IDs.
- * TODO: Implement this function.
  */
 std::unique_ptr<std::vector<type::Value>> StatsStorage::GetColumnStatsByID(
     oid_t database_id, oid_t table_id, oid_t column_id) {
@@ -161,7 +146,7 @@ ResultType StatsStorage::AnalyzeStatsForAllTables() {
       auto table = database->GetTable(table_offset);
       std::unique_ptr<TableStats> table_stats(new TableStats(table));
       table_stats->CollectColumnStats();
-      AddOrUpdateTableStats(table, table_stats.get());
+      InsertOrUpdateTableStats(table, table_stats.get());
     }
   }
   return ResultType::SUCCESS;
@@ -176,7 +161,7 @@ ResultType StatsStorage::AnalyzeStatsForTable(storage::DataTable *table,
   }
   std::unique_ptr<TableStats> table_stats(new TableStats(table));
   table_stats->CollectColumnStats();
-  AddOrUpdateTableStats(table, table_stats.get(), txn);
+  InsertOrUpdateTableStats(table, table_stats.get(), txn);
   return ResultType::SUCCESS;
 }
 
@@ -186,79 +171,6 @@ ResultType StatsStorage::AnalyzeStatsForTable(storage::DataTable *table,
 //               UNUSED_ATTRIBUTE std::vector<std::string> column_names) {
 
 // }
-
-/**
- * CreateSamplesDatabase - Create a database for storing samples tables.
- */
-void StatsStorage::CreateSamplesDatabase() {
-  auto &txn_manager = concurrency::TransactionManagerFactory::GetInstance();
-  auto txn = txn_manager.BeginTransaction();
-  catalog::Catalog::GetInstance()->CreateDatabase(SAMPLES_DB_NAME, txn);
-  txn_manager.CommitTransaction(txn);
-}
-
-/**
- * AddSamplesTable - Add a samples table into the 'samples_db'.
- * The table name is generated by concatenating db_id and table_id with '_'.
- */
-void StatsStorage::AddSamplesTable(
-    storage::DataTable *data_table,
-    std::vector<std::unique_ptr<storage::Tuple>> &sampled_tuples) {
-  auto schema = data_table->GetSchema();
-  auto schema_copy = catalog::Schema::CopySchema(schema);
-  std::unique_ptr<catalog::Schema> schema_ptr(schema_copy);
-  auto catalog = catalog::Catalog::GetInstance();
-  bool is_catalog = false;
-  std::string samples_table_name = GenerateSamplesTableName(
-      data_table->GetDatabaseOid(), data_table->GetOid());
-
-  auto &txn_manager = concurrency::TransactionManagerFactory::GetInstance();
-  auto txn = txn_manager.BeginTransaction();
-  catalog->CreateTable(std::string(SAMPLES_DB_NAME), samples_table_name,
-                       std::move(schema_ptr), txn, is_catalog);
-
-  auto samples_table = catalog->GetTableWithName(std::string(SAMPLES_DB_NAME),
-                                                 samples_table_name, txn);
-
-  for (auto &tuple : sampled_tuples) {
-    InsertSampleTuple(samples_table, std::move(tuple), txn);
-  }
-  txn_manager.CommitTransaction(txn);
-}
-
-bool StatsStorage::InsertSampleTuple(storage::DataTable *samples_table,
-                                     std::unique_ptr<storage::Tuple> tuple,
-                                     concurrency::Transaction *txn) {
-  if (txn == nullptr) {
-    return false;
-  }
-
-  std::unique_ptr<executor::ExecutorContext> context(
-      new executor::ExecutorContext(txn));
-  planner::InsertPlan node(samples_table, std::move(tuple));
-  executor::InsertExecutor executor(&node, context.get());
-  executor.Init();
-  bool status = executor.Execute();
-
-  return status;
-}
-
-/**
- * GetTupleSamples - Query tuple samples by db_id and table_id.
- * Implement this function.
- */
-void StatsStorage::GetTupleSamples(
-    UNUSED_ATTRIBUTE oid_t database_id, UNUSED_ATTRIBUTE oid_t table_id,
-    UNUSED_ATTRIBUTE std::vector<storage::Tuple> &tuple_samples) {}
-
-/**
- * GetColumnSamples - Query column samples by db_id, table_id and column_id.
- * TODO: Implement this function.
- */
-void StatsStorage::GetColumnSamples(
-    UNUSED_ATTRIBUTE oid_t database_id, UNUSED_ATTRIBUTE oid_t table_id,
-    UNUSED_ATTRIBUTE oid_t column_id,
-    UNUSED_ATTRIBUTE std::vector<type::Value> &column_samples) {}
 
 } /* namespace optimizer */
 } /* namespace peloton */
